Minor refactor and lint
